#30 Alpha version of "Kill Team" script and some new functions to support it. (To use it instead of current script rename main_killteam.lua to main.lua

+++  course-work-ps-31-2013 --username dertkaes@gmail.com/OpenGLTest/OpenGLTest/view/GameView.h
@@ -56,8 +56,8 @@ public:
 	void CameraTranslateRight();
 	void CameraTranslateDown();
 	void CameraTranslateUp();
-	void RulerBegin(int x, int y);
-	void RulerEnd(int x, int y);
+	void RulerBegin(double x, double y);
+	void RulerEnd(double x, double y);
 	void RulerHide();
 	void TryMoveSelectedObject(int x, int y);
 	bool UILeftMouseButtonDown(int x, int y);
